feat: update mytimetable
