+++ src/app/talks/me/MyTimetablePage.tsx
@@ -3,8 +3,10 @@
 import {
   AlertTriangle,
   Check,
+  Eye,
   Info,
   List,
+  Pencil,
   QrCode,
   RotateCcw,
   Search,
@@ -14,6 +16,7 @@ import {
 import Link from "next/link";
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
 import { useEffect, useMemo, useRef, useState } from "react";
+import { TimelineColumn } from "@/components/talks/TimelineColumn";
 import {
   DesktopTimelineLayout,
   MobileTimelineLayout,
@@ -25,7 +28,6 @@ import { Input } from "@/ui/input";
 import {
   findOverlaps,
   getTalksByDateFromIds,
-  MY_TIMETABLE_CONST,
   myParticipatedIds,
   myTimetable,
   myTimetableIds,
@@ -42,125 +44,6 @@ type OverlapState = {
   targetTalk: TalkWithMinutes;
   overlaps: TalkWithMinutes[];
 } | null;
-
-function TimelineColumn({
-  eventDate,
-  talks,
-  participatedIds,
-  onClickTimeSlot,
-  onRemoveTalk,
-}: {
-  eventDate: EventDate;
-  talks: TalkWithMinutes[];
-  participatedIds: Set<string>;
-  onClickTimeSlot: (eventDate: EventDate, minutes: number) => void;
-  onRemoveTalk: (id: string) => void;
-}) {
-  const positionedTalks = myTimetable.getPositionedTalks(talks);
-
-  return (
-    <div
-      className="relative rounded-lg border border-black-300 bg-blue-purple-100/30"
-      style={{ height: `${MY_TIMETABLE_CONST.TIMELINE_HEIGHT}px` }}
-    >
-      {MY_TIMETABLE_CONST.TIMELINE_MARKERS.slice(0, -1).map((start) => {
-        const top = myTimetable.minutesToTop(start);
-        const height = myTimetable.minutesToPx(30);
-        return (
-          <button
-            type="button"
-            key={`${eventDate}-${start}`}
-            className="absolute left-0 right-0 z-0 hover:bg-blue-purple-200/40"
-            style={{ top: `${top}px`, height: `${height}px` }}
-            onClick={() => onClickTimeSlot(eventDate, start)}
-            aria-label={`${eventDate} ${myTimetable.formatMinutes(start)} の時間帯で追加`}
-          />
-        );
-      })}
-
-      {MY_TIMETABLE_CONST.TIMELINE_MARKERS.map((minutes) => {
-        const top = myTimetable.minutesToTop(minutes);
-        return (
-          <div
-            key={`${eventDate}-line-${minutes}`}
-            className="absolute left-0 right-0 z-10 border-t border-black-200"
-            style={{ top: `${top}px` }}
-          />
-        );
-      })}
-
-      {positionedTalks.map((talk) => {
-        const top = myTimetable.minutesToTop(talk.startMinutes);
-        const height = myTimetable.minutesToPx(
-          talk.endMinutes - talk.startMinutes,
-        );
-
-        const left = `calc(${(100 / talk.columnCount) * talk.columnIndex}% + 8px)`;
-        const width = `calc(${100 / talk.columnCount}% - 10px)`;
-        const isParticipated = participatedIds.has(talk.id);
-
-        return (
-          <div
-            key={talk.id}
-            className={`absolute left-2 right-2 z-20 overflow-hidden rounded-md border border-black-300 bg-white py-1 px-2 pr-5 border-l-4 ${myTimetable.getTrackBorderClass(talk.track)}`}
-            title={`${talk.time} / ${TRACK[talk.track].name}\n${talk.title}\n${talk.speaker.name}`}
-            style={{
-              top: `${top}px`,
-              height: `${height}px`,
-              left,
-              width,
-              right: "auto",
-            }}
-          >
-            {talk.isOverlapping && (
-              <div
-                className="absolute inset-0 opacity-25 pointer-events-none"
-                style={{
-                  backgroundImage:
-                    "repeating-linear-gradient(-45deg, #ff5a5a 0 6px, transparent 6px 12px)",
-                }}
-              />
-            )}
-            <button
-              type="button"
-              className="absolute right-1 top-1 z-10 cursor-pointer text-black-500 hover:text-black-700"
-              onClick={() => onRemoveTalk(talk.id)}
-              aria-label="削除"
-            >
-              <X size={14} />
-            </button>
-            {isParticipated && (
-              <span className="absolute right-1 bottom-1 z-10 inline-flex items-center gap-0.5 rounded-full bg-green-100 px-1.5 py-0.5 text-[9px] font-bold text-green-700">
-                <Check size={10} />
-                参加済
-              </span>
-            )}
-            <p className="relative z-10 text-[10px] text-black-500 truncate whitespace-nowrap">
-              {talk.time} / {TRACK[talk.track].name}
-            </p>
-            <Link
-              href={`/talks/${talk.id}`}
-              className="relative z-10 hover:underline block"
-            >
-              <p className="mt-0.5 text-xs font-bold text-black-700 truncate">
-                {talk.title}
-              </p>
-            </Link>
-            <p className="relative z-10 mt-0.5 text-[10px] text-black-500 truncate">
-              {talk.speaker.name}
-            </p>
-          </div>
-        );
-      })}
-
-      {talks.length === 0 && (
-        <div className="absolute inset-0 z-10 pointer-events-none flex items-center justify-center text-sm text-black-500">
-          未選択です
-        </div>
-      )}
-    </div>
-  );
-}
 
 function TimePickerDialog({
   timePickerState,
@@ -445,9 +328,13 @@ function InfoDialog({ onClose }: { onClose: () => void }) {
 function SideToolbar({
   selectedIds,
   participatedIds,
+  isViewMode,
+  onToggleViewMode,
 }: {
   selectedIds: string[];
   participatedIds: string[];
+  isViewMode: boolean;
+  onToggleViewMode: () => void;
 }) {
   const [showQr, setShowQr] = useState(false);
   const [isInfoOpen, setIsInfoOpen] = useState(false);
@@ -561,6 +448,16 @@ function SideToolbar({
         >
           <List size={18} />
         </Link>
+      </Button>
+      <Button
+        type="button"
+        variant="outline"
+        size="icon"
+        onClick={onToggleViewMode}
+        aria-label={isViewMode ? "編集モードへ" : "閲覧モードへ"}
+        title={isViewMode ? "編集モードへ" : "閲覧モードへ"}
+      >
+        {isViewMode ? <Pencil size={18} /> : <Eye size={18} />}
       </Button>
       <Button
         type="button"
@@ -718,6 +615,7 @@ export default function MyTimetablePage() {
   const [timePickerState, setTimePickerState] = useState<TimePickerState>(null);
   const [overlapState, setOverlapState] = useState<OverlapState>(null);
   const [isClearConfirmOpen, setIsClearConfirmOpen] = useState(false);
+  const [isViewMode, setIsViewMode] = useState(false);
   const allTalksWithMinutes = useMemo(
     () => myTimetable.getAllTalksWithMinutes(),
     [],
@@ -726,11 +624,6 @@ export default function MyTimetablePage() {
   const talksByDate = useMemo(
     () => getTalksByDateFromIds(selectedIds),
     [selectedIds],
-  );
-
-  const participatedIdsSet = useMemo(
-    () => new Set(participatedIds),
-    [participatedIds],
   );
 
   const hasQueryIds = searchParams.has("m") || searchParams.has("p");
@@ -849,62 +742,100 @@ export default function MyTimetablePage() {
         <SideToolbar
           selectedIds={selectedIds}
           participatedIds={participatedIds}
+          isViewMode={isViewMode}
+          onToggleViewMode={() => setIsViewMode((prev) => !prev)}
         />
 
-        <div className="bg-white rounded-xl p-4 md:p-6">
-          <TalkSearchPanel
-            allTalks={allTalksWithMinutes}
-            selectedIds={selectedIds}
-            hasOpenDialog={!!overlapState}
-            onAdd={addTalk}
-            onRemove={removeTalk}
-            onReset={() => setIsClearConfirmOpen(true)}
-          />
+        {isViewMode ? (
+          <div>
+            <div className="block lg:hidden">
+              <MobileTimelineLayout
+                currentEventDate={currentEventDate}
+                onTabChange={setCurrentEventDate}
+              >
+                <TimelineColumn
+                  eventDate={currentEventDate}
+                  talks={talksByDate[currentEventDate]}
+                  participatedIds={participatedIds}
+                />
+              </MobileTimelineLayout>
+            </div>
 
-          <div className="block lg:hidden">
-            <MobileTimelineLayout
-              currentEventDate={currentEventDate}
-              onTabChange={setCurrentEventDate}
-            >
-              <TimelineColumn
-                eventDate={currentEventDate}
-                talks={talksByDate[currentEventDate]}
-                participatedIds={participatedIdsSet}
-                onClickTimeSlot={(eventDate, minutes) =>
-                  setTimePickerState({ eventDate, minutes })
+            <div className="hidden lg:block">
+              <DesktopTimelineLayout
+                day1Column={
+                  <TimelineColumn
+                    eventDate="DAY1"
+                    talks={talksByDate.DAY1}
+                    participatedIds={participatedIds}
+                  />
                 }
-                onRemoveTalk={removeTalk}
+                day2Column={
+                  <TimelineColumn
+                    eventDate="DAY2"
+                    talks={talksByDate.DAY2}
+                    participatedIds={participatedIds}
+                  />
+                }
               />
-            </MobileTimelineLayout>
+            </div>
           </div>
-
-          <div className="hidden lg:block">
-            <DesktopTimelineLayout
-              day1Column={
-                <TimelineColumn
-                  eventDate="DAY1"
-                  talks={talksByDate.DAY1}
-                  participatedIds={participatedIdsSet}
-                  onClickTimeSlot={(eventDate, minutes) =>
-                    setTimePickerState({ eventDate, minutes })
-                  }
-                  onRemoveTalk={removeTalk}
-                />
-              }
-              day2Column={
-                <TimelineColumn
-                  eventDate="DAY2"
-                  talks={talksByDate.DAY2}
-                  participatedIds={participatedIdsSet}
-                  onClickTimeSlot={(eventDate, minutes) =>
-                    setTimePickerState({ eventDate, minutes })
-                  }
-                  onRemoveTalk={removeTalk}
-                />
-              }
+        ) : (
+          <div className="bg-white rounded-xl p-4 md:p-6">
+            <TalkSearchPanel
+              allTalks={allTalksWithMinutes}
+              selectedIds={selectedIds}
+              hasOpenDialog={!!overlapState}
+              onAdd={addTalk}
+              onRemove={removeTalk}
+              onReset={() => setIsClearConfirmOpen(true)}
             />
+
+            <div className="block lg:hidden">
+              <MobileTimelineLayout
+                currentEventDate={currentEventDate}
+                onTabChange={setCurrentEventDate}
+              >
+                <TimelineColumn
+                  eventDate={currentEventDate}
+                  talks={talksByDate[currentEventDate]}
+                  participatedIds={participatedIds}
+                  onClickTimeSlot={(eventDate, minutes) =>
+                    setTimePickerState({ eventDate, minutes })
+                  }
+                  onRemoveTalk={removeTalk}
+                />
+              </MobileTimelineLayout>
+            </div>
+
+            <div className="hidden lg:block">
+              <DesktopTimelineLayout
+                day1Column={
+                  <TimelineColumn
+                    eventDate="DAY1"
+                    talks={talksByDate.DAY1}
+                    participatedIds={participatedIds}
+                    onClickTimeSlot={(eventDate, minutes) =>
+                      setTimePickerState({ eventDate, minutes })
+                    }
+                    onRemoveTalk={removeTalk}
+                  />
+                }
+                day2Column={
+                  <TimelineColumn
+                    eventDate="DAY2"
+                    talks={talksByDate.DAY2}
+                    participatedIds={participatedIds}
+                    onClickTimeSlot={(eventDate, minutes) =>
+                      setTimePickerState({ eventDate, minutes })
+                    }
+                    onRemoveTalk={removeTalk}
+                  />
+                }
+              />
+            </div>
           </div>
-        </div>
+        )}
       </div>
 
       {timePickerState && (
